% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_termsC}
\alias{get_termsC}
\title{A function helps to get terms using Rcpp}
\usage{
get_termsC(x, vocab)
}
\arguments{
\item{x}{the word in one document}

\item{vocab}{the all word list}
}
\value{
a index
}
\description{
A function helps to get terms using Rcpp
}
