% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAR.R
\name{LDAR}
\alias{LDAR}
\title{Title a function helps for LDA model}
\usage{
LDAR(data, num_topic, num_word)
}
\arguments{
\item{data}{a document-word list}

\item{num_topic}{the number of topics}

\item{num_word}{the number of the top words to return}
}
\value{
a list with top num_word words of every topic
}
\description{
Title a function helps for LDA model
}
