% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDAR.R
\name{LDAVIS}
\alias{LDAVIS}
\title{Title a function helps for LDA model visualization.}
\usage{
LDAVIS(data)
}
\arguments{
\item{data}{a document-word list}
}
\value{
no return
}
\description{
Title a function helps for LDA model visualization.
}
